requirements.txt
confluent-kafka==2.2.0
PyYAML==6.0.1
python-dateutil==2.8.2 